fixup! feat: support the timeSpecification X.509v3 extension

+++ crypto/x509/v3_timespec.c
@@ -136,7 +136,8 @@ static int i2r_OSSL_DAY_TIME(X509V3_EXT_METHOD *method,
         return 0;
     if (dt->minute && !ASN1_INTEGER_get_int64(&s, dt->second))
         return 0;
-    return BIO_printf(out, "%02lld:%02lld:%02lld", h, m, s) > 0;
+    return BIO_printf(out, "%02lld:%02lld:%02lld",
+                      (long long int)h, (long long int)m, (long long int)s) > 0;
 }
 
 static int i2r_OSSL_DAY_TIME_BAND(X509V3_EXT_METHOD *method,
